// src/api/types/planResponse.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * TeaPlus API
 * OpenAPI spec version: 0.0.1
 */

export interface PlanResponse {
    /** Unique identifier of the plan */
    id: string;
    /** Name of the subscription plan */
    title: string;
    /** Description of the subscription plan */
    description: string;
    /** List of features included of the plan */
    features: string[];
    /** Monthly price */
    monthlyPrice: number;
    /** Yearly price */
    yearlyPrice: number;
    /** Indicates whether the plan is featured or promoted */
    isFeatured: boolean;
}